* Small reorganisation in the user documentation. * Machines can be colored (via the "setstyle" command). * Found two memory leaks.
git-svn-id: https://genericcam.svn.sourceforge.net/svnroot/genericcam/trunk@58 fa0ba55f-42f9-4c21-982c-c68e131b72ab

+++ src/machine/Machine.h
@@ -66,8 +66,8 @@ public:
 private:
 	bool initialized;
 
-	int componentWithTool;
-	int componentWithMaterial;
+	size_t componentWithTool;
+	size_t componentWithMaterial;
 	AffineTransformMatrix toolPositionRelativ;
 	AffineTransformMatrix workpiecePositionRelativ;
